Artifact GET working with foreign tables

+++ giflex/src/main/java/br/ufsm/csi/poow2/giflex/repository/CharacterRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Collection;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<Character, Long> {
+public interface CharacterRepository extends JpaRepository<Character, Integer> {
     Collection<? extends Character> findByName(String name);
 }
